# backend/docs/swagger.yaml
basePath: /
definitions:
  dtos.AccountResponse:
    properties:
      balance:
        type: string
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_system:
        type: boolean
      name:
        type: string
      owner_id:
        type: string
    type: object
  dtos.BalanceResponse:
    properties:
      account_id:
        type: string
      calculated_balance:
        type: string
    type: object
  dtos.CreateAccountDto:
    properties:
      currency:
        example: USD
        type: string
      name:
        example: Savings Account
        type: string
    required:
    - currency
    - name
    type: object
  dtos.CreateAccountResponse:
    properties:
      message:
        type: string
      payload:
        $ref: '#/definitions/dtos.AccountResponse'
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  dtos.GetAccountBalanceResponse:
    properties:
      message:
        type: string
      payload:
        $ref: '#/definitions/dtos.BalanceResponse'
      status:
        type: integer
      success:
        type: boolean
    type: object
  dtos.GetAccountResponse:
    properties:
      message:
        type: string
      payload:
        $ref: '#/definitions/dtos.AccountResponse'
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  dtos.ListAccountsResponse:
    properties:
      message:
        type: string
      payload:
        items:
          $ref: '#/definitions/dtos.AccountResponse'
        type: array
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  dtos.LoginResponse:
    properties:
      email:
        type: string
      token:
        type: string
      user_id:
        type: string
    type: object
  dtos.LoginUserDto:
    description: Login credentials for authenticating a user
    properties:
      email:
        description: |-
          User's email address
          @example john.doe@example.com
        example: john.doe@example.com
        type: string
      password:
        description: |-
          User's password
          @example SecureP@ssw0rd
        example: SecureP@ssw0rd
        type: string
    required:
    - email
    - password
    type: object
...
  dtos.RegisterResponse:
    properties:
      email:
        type: string
      token:
        type: string
      user_id:
        type: string
    type: object
  dtos.RegisterUserDto:
    description: Registration data for creating a new user account
    properties:
      email:
        description: |-
          User's email address
          @example john.doe@example.com
        example: john.doe@example.com
        type: string
      password:
        description: |-
          User's password (min 8 characters)
          @example SecureP@ssw0rd
        example: SecureP@ssw0rd
        type: string
    required:
    - email
    - password
    type: object
  dtos.StructuredResponse:
    description: Standard response format containing success status, HTTP status code,
      message, and optional payload
    properties:
      message:
        description: |-
          Human-readable message
          @example Operation completed successfully
        example: Operation completed successfully
        type: string
      payload:
        description: |-
          Optional payload containing response data
          @example {"id":1,"name":"Example Item"}
      status:
        description: |-
          HTTP status code
          @example 200
        example: 200
        type: integer
      success:
        description: |-
          Whether the operation was successful
          @example true
        example: true
        type: boolean
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Double-entry accounting ledger
  title: Double-Entry Bank Ledger API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: Retrieves all accounts belonging to the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ListAccountsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: List user accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Creates a new financial account for the authenticated user.
      parameters:
      - description: Account payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateAccountDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.CreateAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Create a new account
      tags:
      - accounts
...
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a single account belonging to the authenticated user.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GetAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Get an account
      tags:
      - accounts
  /accounts/{id}/balance:
    get:
      consumes:
      - application/json
      description: Retrieves the current balance of a specific account.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.GetAccountBalanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Get account balance
      tags:
      - accounts
  /accounts/{id}/deactivate:
    patch:
      consumes:
      - application/json
      description: Deactivates a user-owned account. This action is irreversible.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Deactivate an account
      tags:
      - accounts
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Validates credentials and returns an access token + sets a refresh
        token cookie.
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.LoginUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      summary: Authenticate a user
      tags:
      - auth
  /auth/logout:
    post:
      description: Revokes the current refresh token and clears the cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      security:
      - BearerAuth: []
      summary: Logout current session
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with the provided credentials.
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.RegisterUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.StructuredResponse'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"